+++ pkg/processor/k8sprocessor/options.go
@@ -20,6 +20,7 @@ import (
 	"regexp"
 	"strings"
 
+	conventions "go.opentelemetry.io/collector/semconv/v1.18.0"
 	"k8s.io/apimachinery/pkg/selection"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8sconfig"
@@ -49,7 +50,9 @@ const (
 	metadataStartTime       = "startTime"
 	metadataStatefulSetName = "statefulSetName"
 
-	deprecatedMetadataClusterName = "clusterName"
+	metadataOtelSemconvServiceName = "k8s.service.name"  // no semantic convention for service name as of right now, but this is reasonable
+	metadataOtelPodStartTime       = "k8s.pod.startTime" // no semantic convention for this, but keeping a similar format for consistency
+	deprecatedMetadataClusterName  = "clusterName"
 )
 
 // Option represents a configuration option that can be passes.
@@ -106,39 +109,39 @@ func WithExtractMetadata(fields ...string) Option {
 		}
 		for _, field := range fields {
 			switch field {
-			case metadataContainerID:
+			case metadataContainerID, conventions.AttributeContainerID:
 				p.rules.ContainerID = true
-			case metadataContainerImage:
+			case metadataContainerImage, conventions.AttributeContainerImageName:
 				p.rules.ContainerImage = true
-			case metadataContainerName:
+			case metadataContainerName, conventions.AttributeContainerName:
 				p.rules.ContainerName = true
-			case metadataCronJobName:
+			case metadataCronJobName, conventions.AttributeK8SCronJobName:
 				p.rules.CronJobName = true
-			case metadataDaemonSetName:
+			case metadataDaemonSetName, conventions.AttributeK8SDaemonSetName:
 				p.rules.DaemonSetName = true
-			case metadataDeploymentName:
+			case metadataDeploymentName, conventions.AttributeK8SDeploymentName:
 				p.rules.DeploymentName = true
-			case metadataHostName:
+			case metadataHostName, conventions.AttributeHostName:
 				p.rules.HostName = true
-			case metadataJobName:
+			case metadataJobName, conventions.AttributeK8SJobName:
 				p.rules.JobName = true
-			case metadataNamespace:
+			case metadataNamespace, conventions.AttributeK8SNamespaceName:
 				p.rules.Namespace = true
-			case metadataNodeName:
+			case metadataNodeName, conventions.AttributeK8SNodeName:
 				p.rules.NodeName = true
-			case metadataPodID:
+			case metadataPodID, conventions.AttributeK8SPodUID:
 				p.rules.PodUID = true
-			case metadataPodName:
+			case metadataPodName, conventions.AttributeK8SPodName:
 				p.rules.PodName = true
-			case metadataReplicaSetName:
+			case metadataReplicaSetName, conventions.AttributeK8SReplicaSetName:
 				p.rules.ReplicaSetName = true
-			case metadataServiceName:
+			case metadataServiceName, metadataOtelSemconvServiceName:
 				p.rules.ServiceName = true
-			case metadataStartTime:
+			case metadataStartTime, metadataOtelPodStartTime:
 				p.rules.StartTime = true
-			case metadataStatefulSetName:
+			case metadataStatefulSetName, conventions.AttributeK8SStatefulSetName:
 				p.rules.StatefulSetName = true
-			case deprecatedMetadataClusterName:
+			case deprecatedMetadataClusterName, conventions.AttributeK8SClusterName:
 				p.logger.Warn("clusterName metadata field has been deprecated and will be removed soon")
 			default:
 				return fmt.Errorf("\"%s\" is not a supported metadata field", field)